backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
Deprecated==1.2.14
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
google-auth==2.36.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib_resources==6.1.2
Jinja2==3.1.3
limits==3.9.0
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
resend==0.7.2
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.0
SQLAlchemy==2.0.26
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
wrapt==1.16.0
